refactor(imports): standardize event bus import patterns across codebase
Convert domain/__init__.py wrapper functions to _LAZY_IMPORTS pattern,
standardize intra-domain imports to use public API, and update test files.

Changes:
- domain/__init__.py: Remove explicit wrapper functions (get_publisher,
  publish, subscribe), add to _LAZY_IMPORTS dict, expose EventHandler,
  add TYPE_CHECKING imports for mypy
- domain/dispatch_coordinator.py: Import publish from vibe3.domain
- domain/handler_registry.py: Import DomainEvent and subscribe from
  vibe3.domain
- tests: Update 3 test files to use public API imports

Verification:
- All affected tests pass (438 tests)
- Modularity tests pass (37 tests)
- Public API verification successful

Issue: #2604
